// types/core.d.ts
export type Direction = 'neutral' | 'up' | 'down' | 'left' | 'right' | 'forward' | 'back' | 'upForward' | 'upBack' | 'downForward' | 'downBack';
export interface PlayerInputMappings {
    up: string;
    down: string;
    left: string;
    right: string;
    lightPunch: string;
    mediumPunch: string;
    heavyPunch: string;
    lightKick: string;
    mediumKick: string;
    heavyKick: string;
}